update: imports & atomic design
