+++ src/components/01-atoms/index.tsx
@@ -1,2 +1,3 @@
-export * from "./TheHeader";
+export * from "./NavbarMobile";
 export * from "./TheFooter";
+export * from "./ui/button";
