+++ src/components/02-molecules/index.tsx
@@ -1,0 +1,1 @@
+export * from "./TheHeader";
