+++ src/components/04-templates/HomeSection.tsx
@@ -1,10 +1,9 @@
-import { TheHeader, TheFooter } from "@/components/01-atoms";
+import { TheHeader } from "@/components/02-molecules";
 
 export const HomeSection = () => {
 	return (
 		<section className="flex w-full flex-col">
 			<TheHeader />
-			<TheFooter />
 		</section>
 	);
 };
